+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Optional command line parameter `-ld` for specifying source library directory.
+- Optimisation `r` for removing REM commands from source code.
 
 ### Changed